Replace Select with SelectComponent and add PopOver
Updated the import statement to use SelectComponent instead of Select.
Introduced a new PopOver component and imported it from "@/components/PopOver".
Added an options array containing fruit and vegetable options for the SelectComponent.
Introduced a state variable selectedValue to manage the selected option.
Replaced the Select component in the Modal with SelectComponent, using the options array and setting a default value.
Added a new section titled "Popover Example" with a PopOver component providing additional information.